requirements.txt
lightning
matplotlib
PyYAML
einops
tqdm
tensorboard
scipy
ipywidgets